Don't suppress logs even when passing sensitive ssh private key
When passing in a value marked as sensitive for the `ssh_private_key`
the logs are redacted as they should be, except we don't want the logs
of the nixos deploy to be redacted because we want feedback about what
is happening.

We mark the `ssh_private_key` as sensitive to make clear that it should
be sensitive. We then discard the sensitivity when running the
`nixos-deploy.sh` script and mark the `local-exec` provisioner to be
quiet so that terraform does not log the sensitive variable which is
interpolated into the interpreter string.